+++ src/lib/common/XmppQueue/XmppClientWrapper.ts
@@ -44,6 +44,7 @@ const _resourceBinding = require('@xmpp/resource-binding').default || require('@
 const plain = require('@xmpp/sasl-plain');
 // */
 const scramsha512 = require('./sasl-scram-sha-512');
+const scramsha256 = require('./sasl-scram-sha-256');
 const scramsha1 = require("@xmpp/sasl-scram-sha-1").default || require("@xmpp/sasl-scram-sha-1");
 const plain = require('@xmpp/sasl-plain').default || require('@xmpp/sasl-plain');
 //const anonymous = require('@xmpp/sasl-anonymous');
@@ -82,17 +83,20 @@ function client(options = {}) {
 
   const saslFactory = new SASLFactory();
   // SASL mechanisms - order matters and define priority
-  const mechanisms = Object.entries({ scramsha512, scramsha1, plain}).map(
+  const mechanisms = Object.entries({ scramsha512, scramsha256, scramsha1, plain}).map(
   //const mechanisms = Object.entries({ plain}).map(
     ([k, v]) => ({[k]: v(saslFactory)})
   );
 
   const sasl = _sasl({streamFeatures, saslFactory}, async (authenticate, mechanisms) => {
     const mechanism = mechanisms.includes('SCRAM-SHA-512') ? 'SCRAM-SHA-512' :
+                      mechanisms.includes('SCRAM-SHA-256') ? 'SCRAM-SHA-256' :
                       mechanisms.includes('SCRAM-SHA-1') ? 'SCRAM-SHA-1' : mechanisms[0];
     await authenticate(credentials || {username, password}, mechanism);
   });
-  console.log("resourceBinding, iqCaller : ", iqCaller, ", streamFeatures :", streamFeatures, ", resource : ", resource)
+
+  console.log("mechanisms : ", mechanisms, " - resourceBinding, iqCaller : ", iqCaller, ", streamFeatures :", streamFeatures, ", resource : ", resource);
+
   const resourceBinding = _resourceBinding({iqCaller, streamFeatures}, resource);
   //const sessionEstablishment = _sessionEstablishment({iqCaller, streamFeatures});